+++ src/parser/inline/autolink/uri.ts
@@ -1,8 +1,7 @@
 ﻿import { AutolinkParser } from '../../inline';
-import { union, some, surround, subline, focus, rewrite, convert } from '../../../combinator';
+import { union, some, surround, subline, rewrite, convert } from '../../../combinator';
 import { unescsource } from '../../source/unescapable';
 import { link, bracket } from '../link';
-import { defrag } from '../../util';
 
 const closer = /^[-+*~^,.;:!?]*(?=[\s"`|\[\](){}<>]|\\?(?:\s|$))/;
 
@@ -15,26 +14,15 @@ export const uri: AutolinkParser.UriParser = subline(union([
         source => `[]{${address(source)}${attribute(source)}}`,
         link)),
     ''),
-  surround(
-    /^!(?=h?ttps?:\/\/[^/?#\s])/,
-    rewrite(
-      some(union([bracket, some(unescsource, closer)])),
-      convert(
-        source => `[![]{${address(source)}}]{${address(source)}${attribute(source)}}`,
-        link)),
-    ''),
-  focus(
-    /^[0-9a-zA-Z!?][!?]*h?ttps?(?=:)/,
-    defrag(some(unescsource))),
 ]));
 
-function address(source: string): string {
+export function address(source: string): string {
   return source.startsWith('ttp')
     ? `h${source}`
     : source;
 }
 
-function attribute(source: string): string {
+export function attribute(source: string): string {
   return source.startsWith('ttp')
     ? ' nofollow'
     : '';